+++ R/ArrayParameter-proto.R
@@ -136,7 +136,7 @@ ArrayParameter <- pproto(
     pkg <- strsplit(self$widget, "::")[[1]][[1]]
     if (!requireNamespace(pkg, quietly = TRUE))
       stop(paste0("the \"", pkg, "\" R package must be installed to render",
-                  " this parameter."))
+                  " this parameter"))
     # extract function
     f <- do.call(utils::getFromNamespace,
       as.list(rev(strsplit(self$widget, "::")[[1]])))